Resolve "dns" Module Externalization Warning in useStrapiClient.mjs
This pull request addresses the warning related to the "dns" module externalization for browser compatibility in the useStrapiClient.mjs file. The warning appeared during the Vite resolve process, and this change ensures that the "dns" module is imported conditionally only in a server environment during development. This modification aligns with best practices for handling browser-incompatible modules and resolves the warning mentioned in the Vite troubleshooting guide.

Please review and merge at your earliest convenience. If there are any specific guidelines or adjustments needed, feel free to provide feedback. Thank you!